Github setup-gdal action initial commit
Github workflow action to build and install the python OSGeo/gdal
module in the current job. An intermediate pip wheel is generated, that
can can be exported by "build-wheel.yml" workflow.

Under "windows" runners, it downloads the GIS SDK, instead of doing a
fresh build. The suppoort for "macos" runner is not yes finished.