Fix the disabled libkml feature
The "cmake configure" output contains the message "Could NOT find LibKML
(missing: LIBKML_INCLUDE_DIR)", even if "libkml" is added to
"inputs.extra-packages".

This is because of the broken LibKML version in conda-forge channel.
The cmake files are missing here:
  conda-forge/win-64::libkml-1.3.0-h538826c_1021
But are present here
  pkgs/main/win-64::libkml-1.3.0-h63940dd_7

Workaround: Add "defaults::" channel prefix in package name. This was
added to the "extra-packages.default" value in build-nuget.yml.